Fix dependency vulnerabilities
The version of eslint we're depending on has a known security
vulnerability, and although it's just a devDependency I see no good
reason not to fix it.

This commit updates to the latest version, but also relaxes the version
lock to allow PATCH and MINOR version changes in the future.